requirements.txt
pandas>=2.0.0
imap-tools>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
googlemaps>=4.10.0
openpyxl>=3.0.0
pyyaml>=6.0.0

